Refactor default expectations construct
All expectations objects should be lists expressing lines
of python code which are fed into eval().

Previously it was assumed that there may be need to
override other expectations based on a global vs. local
heirarchy. This heirarchy has proven to be redundant since
the endpoint definition expectations are already considered
to be global (highest in heirarchy), followed by the
default expectations (lowest in heirarchy). Therefore, the
expectations dictionary has been simplified to be a simple
list without the need to define what is global or local.